+++ src/bot.js
@@ -4,27 +4,11 @@
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
 const fs = require("fs");
-const Discord = require("discord.js");
+const WelcomeBot = require("./WelcomeBot");
 const dotenv = require("dotenv").config();
-const package = require("../package.json");
 
-const client = new Discord.Client();
+const client = new WelcomeBot();
 
-client.commands = new Discord.Collection();
-client.disabled = new Discord.Collection();
-client.cooldowns = new Discord.Collection();
-client.defaultPrefix = "w/";
-client.categories = [
-    { name: "Setup", emoji: "" },
-    { name: "General", emoji: "" },
-    { name: "Information", emoji: "" },
-    { name: "Moderation", emoji: "" },
-    { name: "Miscellaneous", emoji: "" },
-    { name: "Fun", emoji: "<:fun:854002049095303188>" },
-    { name: "Owner Only", emoji: "" },
-];
-client.botVersion = package.version;
-client.changelog = package.changelog;
 process.env.userAgent = "Discord Welcome-Bot " + client.botVersion;
 process.env.ownerIDs = [
     "815204465937481749" /*PuneetGopinath#6300*/,
@@ -33,48 +17,6 @@ process.env.ownerIDs = [
 process.on("exit", (code) => {
     client.destroy();
 });
-
-const commandFolder = __dirname + "/commands";
-const commandFolders = fs.readdirSync(commandFolder);
-let defaultOpts = {
-    bot_perms: [
-        "VIEW_CHANNEL",
-        "SEND_MESSAGES",
-        "READ_MESSAGE_HISTORY",
-        "MANAGE_MESSAGES",
-    ],
-    args: false,
-    catchError: true,
-    disabled: false,
-    cooldown: 3,
-    ownerOnly: false,
-    category: "General",
-};
-
-for (const folder of commandFolders) {
-    const commandFiles = fs
-        .readdirSync(`${commandFolder}/${folder}`)
-        .filter((file) => file.endsWith(".js"));
-    for (const file of commandFiles) {
-        let module = file.replace(".js", "");
-        let command = require(`${commandFolder}/${folder}/${module}`);
-        if (command.bot_perms) {
-            command.bot_perms = [
-                ...defaultOpts.bot_perms,
-                ...command.bot_perms,
-            ];
-        }
-        command = {
-            ...defaultOpts,
-            ...command,
-        };
-        if (!command.disabled) {
-            client.commands.set(command.name, command);
-        } else {
-            client.disabled.set(command.name, command);
-        }
-    }
-}
 
 const presence = require("./functions/presence");
 const greetUser = require("./functions/greetUser");
